Added heterogeneous chains: infixl1, infixr1, prefix, and postfix (Pattern 1b: Heterogeneous Chains)
```
{---------------------------------------------}
{-     Pattern 1b: Heterogeneous Chains      -}
{---------------------------------------------}
{- For associative binary operators where    -}
{- operand types may differ, use `infixl1`   -}
{- or `infixr1` to combine operands with     -}
{- their operators, in conjunction with      -}
{- strongly typed semantic actions           -}
{---------------------------------------------}
```

The problem with the traditional `chainl1` and `chainr1` is that they are indistinguishable
in their types, which can cause minor typos to cause bugs. We offer _refined_ versions called
`infixl1` and infixr1` that encode the associativity in the types.

To use these, the AST needs to more strongly adhere to the grammar, though the original chains
can be recovered by using `id :: a -> a`.