Used OverloadedStrings to encapsulate lexing behaviour, tests working! (Pattern 2c: Overloaded Strings)
```
{---------------------------------------------}
{-       Pattern 2c: Overloaded Strings      -}
{---------------------------------------------}
{- Hide tokenising logic by allowing string  -}
{- literals to serve as parsers.             -}
{---------------------------------------------}
```

By using `OverloadedStrings` we can hide all the tokenising
logic behind the `IsString` instance for `Parser ()`. This means
that we can add and remove keywords without needing to change
anything in the parser, whilst keeping it looking clean. If
the extension is undesirable, a combinator performing the job
of `fromString` is also a great idea.